+++ src/clj/backtype/storm/disruptor.clj
@@ -46,8 +46,14 @@
 (defmacro handler [& args]
   `(clojure-handler (fn ~@args)))
 
-(defn publish [^DisruptorQueue q o]
-  (.publish q o))
+(defn publish
+  ([^DisruptorQueue q o block?]
+    (.publish q o block?))
+  ([q o]
+    (publish q o true)))
+
+(defn try-publish [^DisruptorQueue q o]
+  (.tryPublish q o))
 
 (defn consume-batch [^DisruptorQueue queue handler]
   (.consumeBatch queue handler))