+++ src/jvm/backtype/storm/scheduler/SchedulerAssignmentImpl.java
@@ -3,6 +3,7 @@ package backtype.storm.scheduler;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -24,6 +25,11 @@ public class SchedulerAssignmentImpl implements SchedulerAssignment {
             this.executorToSlot.putAll(executorToSlots);
         }
     }
+
+    @Override
+    public Set<WorkerSlot> getSlots() {
+        return new HashSet(executorToSlot.values());
+    }    
     
     /**
      * Assign the slot to executors.